OBGM-47 Refactor UserlocationExcelImporter
